Refactor python cssbeautifier to reuse jsbeautifier CLI methods
This change also adds smoke tests for installed cssbeautifier
package.
