+++ python/jsbeautifier/__init__.py
@@ -10,6 +10,7 @@ import errno
 import copy
 import glob
 from jsbeautifier.__version__ import __version__
+from jsbeautifier.cli import *
 from jsbeautifier.javascript.options import BeautifierOptions
 from jsbeautifier.javascript.beautifier import Beautifier
 
@@ -63,9 +64,13 @@ from jsbeautifier.javascript.beautifier import Beautifier
 #
 # Here are the available options: (read source)
 
-
-class MissingInputStreamError(Exception):
-    pass
+__all__ = [
+    "default_options",
+    "beautify",
+    "beautify_file",
+    "usage",
+    "main",
+]
 
 
 def default_options():
@@ -77,75 +82,8 @@ def beautify(string, opts=default_options()):
     return b.beautify(string, opts)
 
 
-def set_file_editorconfig_opts(filename, js_options):
-    from editorconfig import get_properties, EditorConfigError
-
-    try:
-        _ecoptions = get_properties(os.path.abspath(filename))
-
-        if _ecoptions.get("indent_style") == "tab":
-            js_options.indent_with_tabs = True
-        elif _ecoptions.get("indent_style") == "space":
-            js_options.indent_with_tabs = False
-
-        if _ecoptions.get("indent_size"):
-            js_options.indent_size = int(_ecoptions["indent_size"])
-
-        if _ecoptions.get("max_line_length"):
-            if _ecoptions.get("max_line_length") == "off":
-                js_options.wrap_line_length = 0
-            else:
-                js_options.wrap_line_length = int(_ecoptions["max_line_length"])
-
-        if _ecoptions.get("insert_final_newline") == "true":
-            js_options.end_with_newline = True
-        elif _ecoptions.get("insert_final_newline") == "false":
-            js_options.end_with_newline = False
-
-        if _ecoptions.get("end_of_line"):
-            if _ecoptions["end_of_line"] == "cr":
-                js_options.eol = "\r"
-            elif _ecoptions["end_of_line"] == "lf":
-                js_options.eol = "\n"
-            elif _ecoptions["end_of_line"] == "crlf":
-                js_options.eol = "\r\n"
-
-    except EditorConfigError:
-        # do not error on bad editor config
-        print("Error loading EditorConfig.  Ignoring.", file=sys.stderr)
-
-
 def beautify_file(file_name, opts=default_options()):
-    input_string = ""
-    if file_name == "-":  # stdin
-        if sys.stdin.isatty():
-            raise MissingInputStreamError()
-
-        stream = sys.stdin
-        if platform.platform().lower().startswith("windows"):
-            if sys.version_info.major >= 3:
-                # for python 3 on windows this prevents conversion
-                stream = io.TextIOWrapper(sys.stdin.buffer, newline="")
-            elif platform.architecture()[0] == "32bit":
-                # for python 2 x86 on windows this prevents conversion
-                import msvcrt
-
-                msvcrt.setmode(sys.stdin.fileno(), os.O_BINARY)
-            else:
-                raise Exception(
-                    "Pipe to stdin not supported on Windows with Python 2.x 64-bit."
-                )
-
-        input_string = stream.read()
-
-        # if you pipe an empty string, that is a failure
-        if input_string == "":
-            raise MissingInputStreamError()
-    else:
-        stream = io.open(file_name, "rt", newline="", encoding="UTF-8")
-        input_string = stream.read()
-
-    return beautify(input_string, opts)
+    return process_file(file_name, opts, beautify)
 
 
 def usage(stream=sys.stdout):
@@ -215,24 +153,6 @@ Rarely needed options:
         return 1
     else:
         return 0
-
-
-def mkdir_p(path):
-    try:
-        if path:
-            os.makedirs(path)
-    except OSError as exc:  # Python >2.5
-        if exc.errno == errno.EEXIST and os.path.isdir(path):
-            pass
-        else:
-            raise Exception()
-
-
-def isFileDifferent(filepath, expected):
-    try:
-        return "".join(io.open(filepath, "rt", newline="").readlines()) != expected
-    except BaseException:
-        return True
 
 
 def main():
@@ -357,108 +277,20 @@ def main():
             return usage()
 
     try:
-        filepaths = []
-        if not filepath_params or (
-            len(filepath_params) == 1 and filepath_params[0] == "-"
-        ):
-            # default to stdin
-            filepath_params = []
-            filepaths.append("-")
-
-        for filepath_param in filepath_params:
-            # ignore stdin setting if files are specified
-            if "-" == filepath_param:
-                continue
-
-            # Check if each literal filepath exists
-            if os.path.isfile(filepath_param):
-                filepaths.append(filepath_param)
-            elif "*" in filepath_param or "?" in filepath_param:
-                # handle globs
-                # empty result is okay
-                if sys.version_info.major == 2 or (
-                    sys.version_info.major == 3 and sys.version_info.minor <= 4
-                ):
-                    if "**" in filepath_param:
-                        raise Exception(
-                            "Recursive globs not supported on Python <= 3.4."
-                        )
-                    filepaths.extend(glob.glob(filepath_param))
-                else:
-                    filepaths.extend(glob.glob(filepath_param, recursive=True))
-            else:
-                # not a glob and not a file
-                raise OSError(errno.ENOENT, os.strerror(errno.ENOENT), filepath_param)
-
-        if len(filepaths) > 1:
-            replace = True
-        elif filepaths and filepaths[0] == "-":
-            replace = False
-
-        # remove duplicates
-        filepaths = set(filepaths)
-
+        filepaths, replace = get_filepaths_from_params(filepath_params, replace)
         for filepath in filepaths:
             if not replace:
                 outfile = outfile_param
             else:
                 outfile = filepath
 
-            # Editorconfig used only on files, not stdin
-            if getattr(js_options, "editorconfig"):
-                editorconfig_filepath = filepath
-
-                if editorconfig_filepath == "-":
-                    if outfile != "stdout":
-                        editorconfig_filepath = outfile
-                    else:
-                        fileType = "js"
-                        editorconfig_filepath = "stdin." + fileType
-
-                # debug("EditorConfig is enabled for ", editorconfig_filepath);
-                js_options = copy.copy(js_options)
-                set_file_editorconfig_opts(editorconfig_filepath, js_options)
+            js_options = integrate_editorconfig_options(
+                filepath, js_options, outfile, "js"
+            )
 
             pretty = beautify_file(filepath, js_options)
 
-            if outfile == "stdout":
-                stream = sys.stdout
-
-                # python automatically converts newlines in text to "\r\n" when on windows
-                # switch to binary to prevent this
-                if platform.platform().lower().startswith("windows"):
-                    if sys.version_info.major >= 3:
-                        # for python 3 on windows this prevents conversion
-                        stream = io.TextIOWrapper(sys.stdout.buffer, newline="")
-                    elif platform.architecture()[0] == "32bit":
-                        # for python 2 x86 on windows this prevents conversion
-                        import msvcrt
-
-                        msvcrt.setmode(sys.stdout.fileno(), os.O_BINARY)
-                    else:
-                        raise Exception(
-                            "Pipe to stdout not supported on Windows with Python 2.x 64-bit."
-                        )
-
-                stream.write(pretty)
-            else:
-                if isFileDifferent(outfile, pretty):
-                    mkdir_p(os.path.dirname(outfile))
-
-                    # python automatically converts newlines in text to "\r\n" when on windows
-                    # set newline to empty to prevent this
-                    with io.open(outfile, "wt", newline="", encoding="UTF-8") as f:
-                        print("beautified " + outfile, file=sys.stdout)
-                        try:
-                            f.write(pretty)
-                        except TypeError:
-                            # This is not pretty, but given how we did the version import
-                            # it is the only way to do this without having setup.py
-                            # fail on a missing six dependency.
-                            six = __import__("six")
-                            f.write(six.u(pretty))
-                elif not js_options.keep_quiet:
-                    print("beautified " + outfile + " - unchanged", file=sys.stdout)
+            write_beautified_output(pretty, js_options, outfile)
 
     except MissingInputStreamError:
         print("Must pipe input or define at least one file.\n", file=sys.stderr)
